pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sabre"
version = "0.1.0"
description = "Toolkit for configuring and running adversarial evaluation tournaments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "jsonschema",
    "typer>=0.19.1",
    "rich",
    "rapidfuzz",
    "openai>=1.108.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
sabre = "sabre.cli:app"
saber = "sabre.cli:app"  # legacy shim

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
